+++ CHANGELOG.md
@@ -1,4 +1,13 @@
 changelog
 =============
+* 2017-11-19 21:28
 
-* 2017-11-18 18:49 init 
+    1. add more log;
+    2. fix logout twice will throw null pointer exception;
+    3. change post request to all request except signIn option;
+    4. reduce the length of ticketName and ticketValue;
+    5. log output to file
+    
+* 2017-11-18 18:49
+    
+    1. init;